Create Question-3
This C program generates the Fibonacci sequence up to the nth term, where the user specifies the value of n. The Fibonacci sequence is a series of numbers in which each number is the sum of the two preceding ones, starting with 0 and 1. The program begins by prompting the user to enter the number of terms they want in the sequence. Upon receiving the input, it invokes the Fibonacci function. Inside this function, the first two terms of the sequence are initialized as 0 and 1. Using a loop that iterates 'n' times, each term is calculated by adding the preceding two terms. The generated sequence is then printed to the console. Finally, the program terminates, having successfully displayed the Fibonacci sequence up to the specified term. This program offers a simple and efficient way to understand and visualize the Fibonacci sequence in action.