Create Question-5
This program prompts the user to input a string, then it iterates through the characters of the string to compare characters from the beginning and end simultaneously. If any pair of characters doesn't match, it concludes that the string is not a palindrome. Otherwise, it determines that the string is a palindrome.