Create Question-8
Linear search is a simple searching algorithm that sequentially checks each element of the array until it finds the desired element or reaches the end of the array. This program demonstrates how linear search can be implemented in C to search for a given integer element in an array entered by the user.